--- web/apps/photos/src/components/pages/gallery/AvatarGroup.tsx
@@ -1,52 +1,0 @@
-import type { Collection } from "@/media/collection";
-import { styled } from "@mui/material";
-import NumberAvatar from "@mui/material/Avatar";
-import Avatar from "./Avatar";
-
-const AvatarContainer = styled("div")({
-    position: "relative",
-    display: "flex",
-    alignItems: "center",
-    marginLeft: -5,
-});
-
-const AvatarContainerOuter = styled("div")({
-    position: "relative",
-    display: "flex",
-    alignItems: "center",
-    marginLeft: 8,
-});
-const AvatarCounter = styled(NumberAvatar)({
-    height: 20,
-    width: 20,
-    fontSize: 10,
-    color: "#fff",
-});
-
-const SHAREE_AVATAR_LIMIT = 6;
-
-const AvatarGroup = ({ sharees }: { sharees: Collection["sharees"] }) => {
-    const hasShareesOverLimit = sharees?.length > SHAREE_AVATAR_LIMIT;
-    const countOfShareesOverLimit = sharees?.length - SHAREE_AVATAR_LIMIT;
-
-    return (
-        <AvatarContainerOuter>
-            {sharees?.slice(0, 6).map((sharee) => (
-                <AvatarContainer key={sharee.email}>
-                    <Avatar
-                        key={sharee.email}
-                        email={sharee.email}
-                        opacity={100}
-                    />
-                </AvatarContainer>
-            ))}
-            {hasShareesOverLimit && (
-                <AvatarContainer key="extra-count">
-                    <AvatarCounter>+{countOfShareesOverLimit}</AvatarCounter>
-                </AvatarContainer>
-            )}
-        </AvatarContainerOuter>
-    );
-};
-
-export default AvatarGroup;
